Move SamplesCodeQualityIntegrationTest to :code-quality
